tests/wpt/tests/layout-instability/partially-clipped-visual-rect.html

<!DOCTYPE html>
<title>Layout Instability: partially clipped visual rect</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<style>

body { margin: 0; }
#clip { width: 150px; height: 600px; overflow: hidden; }
#j { position: relative; width: 300px; height: 200px; background: blue; }

</style>
<div id='clip'><div id='j'></div></div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<script>

promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  document.querySelector("#j").style.top = "200px";

  // (clipped width 150px) * (height 200 + movement 200)
  const expectedScore = computeExpectedScore(150 * (200 + 200), 200);

  await watcher.promise;
  assert_equals(watcher.score, expectedScore);
}, "Partially clipped visual rect.");

</script>
